Refactor format -> ml_croissant
