+++ .github/workflows/ci.yml
@@ -20,7 +20,7 @@ jobs:
     runs-on: ubuntu-latest
     defaults:
       run:
-        working-directory: ./python/format
+        working-directory: ./python/ml_croissant
     steps:
     - uses: actions/checkout@v3
 
@@ -43,7 +43,7 @@ jobs:
     runs-on: ubuntu-latest
     defaults:
       run:
-        working-directory: ./python/format
+        working-directory: ./python/ml_croissant
     steps:
     - uses: actions/checkout@v3
 
@@ -73,7 +73,7 @@ jobs:
     runs-on: ubuntu-latest
     defaults:
       run:
-        working-directory: ./python/format
+        working-directory: ./python/ml_croissant
     steps:
     - uses: actions/checkout@v3
