+++ python/ml_croissant/__init__.py
@@ -1,0 +1,2 @@
+from ml_croissant._src.datasets import Dataset
+from ml_croissant._src.errors import ValidationError
