+++ python/ml_croissant/_src/computations_test.py
@@ -1,15 +1,15 @@
 """computations_test module."""
 
 from etils import epath
-from format._src.computations import (
+from ml_croissant._src.computations import (
     Download,
     InitOperation,
     ReadCsv,
     ReadField,
     GroupRecordSet,
 )
-from format._src.errors import Issues
-from format._src.nodes import Node
+from ml_croissant._src.errors import Issues
+from ml_croissant._src.nodes import Node
 import pytest
 import rdflib
 from rdflib import namespace
