+++ python/ml_croissant/_src/constants.py
