+++ python/ml_croissant/scripts/validate.py
@@ -5,8 +5,7 @@ import sys
 from absl import app
 from absl import flags
 from absl import logging
-from format import Dataset
-from format._src import errors
+from ml_croissant import Dataset, ValidationError
 
 flags.DEFINE_string(
     "file",
@@ -26,7 +25,7 @@ def main(argv):
     try:
         Dataset(file)
         logging.info("Done.")
-    except errors.ValidationError as exception:
+    except ValidationError as exception:
         logging.error(exception)
         sys.exit(1)
